Feature: Se ha aplicado 3 modadlides, elite,abc,mentor, ademas del panel de configuracioń de las figuras , en ello generamos las piezas y podemos generar el archivo que nos sirve para hacer el data para poder generarlo
