+++ src/BarcoMENTOR.js
@@ -1,0 +1,72 @@
+export const triangulo = [
+  {
+    id: "p1",
+    nombre: "Triángulo Grande 1",
+    points: [0, 0, 200, 0, 0, 200],
+    x: 306,
+    y: 251,
+    fill: "skyblue",
+    rotation: 180,
+    draggable: true,
+  },
+  {
+    id: "p2",
+    nombre: "Triángulo Grande 2",
+    points: [0, 0, 200, 0, 0, 200],
+    x: 329,
+    y: 250,
+    fill: "orange",
+    rotation: -91,
+    draggable: true,
+  },
+  {
+    id: "p3",
+    nombre: "Triángulo Mediano",
+    points: [0, 0, 140, 0, 0, 140],
+    x: 150,
+    y: 364,
+    fill: "magenta",
+    rotation: -136,
+    draggable: true,
+  },
+  {
+    id: "p4",
+    nombre: "Triángulo Pequeño 1",
+    points: [0, 0, 100, 0, 0, 100],
+    x: 372,
+    y: 263,
+    fill: "lime",
+    rotation: 1,
+    draggable: true,
+  },
+  {
+    id: "p5",
+    nombre: "Triángulo Pequeño 2",
+    points: [0, 0, 100, 0, 0, 100],
+    x: 258,
+    y: 363,
+    fill: "yellow",
+    rotation: 179,
+    draggable: true,
+  },
+  {
+    id: "p6",
+    nombre: "Cuadrado",
+    points: [0, 0, 100, 0, 100, 100, 0, 100],
+    x: 264,
+    y: 262,
+    fill: "blue",
+    rotation: 0,
+    draggable: true,
+  },
+  {
+    id: "p7",
+    nombre: "Paralelogramo",
+    points: [-85, -10, 55, -10, 125, 60, -20, 60],
+    x: 514,
+    y: 317,
+    fill: "teal",
+    rotation: 134,
+    draggable: true,
+  },
+];
