+++ src/CamelloELITE.js
@@ -1,0 +1,72 @@
+export const triangulo = [
+  {
+    id: "p1",
+    nombre: "Triángulo Grande 1",
+    points: [0, 0, 200, 0, 0, 200],
+    x: 421,
+    y: 456,
+    fill: "skyblue",
+    rotation: 135,
+    draggable: true,
+  },
+  {
+    id: "p2",
+    nombre: "Triángulo Grande 2",
+    points: [0, 0, 200, 0, 0, 200],
+    x: 555,
+    y: 386,
+    fill: "orange",
+    rotation: 91,
+    draggable: true,
+  },
+  {
+    id: "p3",
+    nombre: "Triángulo Mediano",
+    points: [0, 0, 140, 0, 0, 140],
+    x: 451,
+    y: 287,
+    fill: "magenta",
+    rotation: 45,
+    draggable: true,
+  },
+  {
+    id: "p4",
+    nombre: "Triángulo Pequeño 1",
+    points: [0, 0, 100, 0, 0, 100],
+    x: 209,
+    y: 238,
+    fill: "lime",
+    rotation: -135,
+    draggable: true,
+  },
+  {
+    id: "p5",
+    nombre: "Triángulo Pequeño 2",
+    points: [0, 0, 100, 0, 0, 100],
+    x: 207,
+    y: 383,
+    fill: "yellow",
+    rotation: -45,
+    draggable: true,
+  },
+  {
+    id: "p6",
+    nombre: "Cuadrado",
+    points: [0, 0, 100, 0, 100, 100, 0, 100],
+    x: 351,
+    y: 243,
+    fill: "blue",
+    rotation: 45,
+    draggable: true,
+  },
+  {
+    id: "p7",
+    nombre: "Paralelogramo",
+    points: [-85, -10, 55, -10, 125, 60, -20, 60],
+    x: 220,
+    y: 293,
+    fill: "teal",
+    rotation: -90,
+    draggable: true,
+  },
+];
